Implement a concurrent queue based on semaphores from concurrent-ruby
A custom queue is necessary because ruby's stdlib queue does not support
a peek operation. This operation is needed to implement the MSG_PEEK
recvfrom flag.

This adds a dependency on `concurrent-ruby` since semaphores are not
included in the ruby standard library.

This queue replaces the queue implementation previously used in the
UDPSocket class for communication with the background thread.